requirements.txt
flask==2.0.1
Flask-SSLify==0.1.5
bcrypt==3.2.0
google-cloud-datastore==2.1.3
google-cloud-storage==1.39.0
google-api-python-client==2.13.0
google-cloud-vision==2.4.0
Pillow==8.3.1
numpy==1.19.5
